Return diff from edit tool
